requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-groq
langchain-chroma
pypdf
fpdf2
python-dotenv
google-generativeai
groq
pillow
requests
markdown
